+++ src/boxyard/cmds/_get_box_sync_status.py
@@ -29,7 +29,7 @@ async def get_box_sync_status(
     from boxyard import const as _const
     
     _conf_path = box_meta.get_local_part_path(config, BoxPart.CONF)
-    _box_exclude_path = _conf_path / ".rclone_exclude"
+    _box_exclude_path = _conf_path / _const.RCLONE_EXCLUDE_FILENAME
     _effective_exclude_path = (
         _box_exclude_path
         if _box_exclude_path.exists()